requirements.txt
telebot
fastai
torch
Pillow
pathlib